pyproject.toml
[project]
name = "warrenbotfett"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "autoflake>=2.3.1",
    "datamodel-code-generator>=0.31.2",
    "diskcache>=5.6.3",
    "dotenv>=0.9.9",
    "google-genai>=1.20.0",
    "isort>=6.0.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.26",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.5.0",
    "pydantic-ai[logfire]>=0.3.4",
    "pyright>=1.1.402",
    "requests>=2.32.4",
    "ruff>=0.12.1",
    "supabase>=2.16.0",
]


[tool.setuptools]
packages = ["compass"]
